--- src/main/java/org/example/namelesschamber/metrics/scheduler/MetricsScheduler.java
@@ -1,26 +1,0 @@
-package org.example.namelesschamber.metrics.scheduler;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.example.namelesschamber.metrics.service.MetricsService;
-import org.example.namelesschamber.common.util.DiscordNotifier;
-import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Component;
-
-import java.util.Map;
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class MetricsScheduler {
-
-    private final MetricsService metricsService;
-    private final DiscordNotifier discordNotifier;
-
-    @Scheduled(cron = "0 0 23 * * *")
-    public void reportMetricsEveryDay() {
-        Map<String, Object> report = metricsService.buildMetricsEmbed();
-        discordNotifier.sendEmbed(report);
-        log.info("Metrics report sent to Discord: {}", report);
-    }
-}
