+++ src/main/resources/application.yml
@@ -33,7 +33,7 @@ security:
       - /auth/signup
       - /auth/anonymous
       - /auth/reissue
-      - /metrics/**
+      - /notifications/metrics/today
 
 discord:
   webhook:
